+++ condow_core/CHANGELOG.md
@@ -10,14 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **BREAKING** Downloading is done via a builder style request object
 - `Reporter` is used dynamically and now called `Probe`
+- **BREAKING** Trait `Downloads` uses request API and has location as associated type
 
 ### ADDED
 
 - support of `tracing crate`
+- trait `DownloadsUntyped` which takes a `&str` as a location
 
 ### REMOVED
 
-- **BREAKING** Removed trait `Downloads`
 - **BREAKING** Old logging mechanism via reporter 
 - **BREAKING** `CompositeReporter`